backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
bcrypt==4.0.1
passlib==1.7.4